% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWeather.R
\docType{package}
\name{ShinyWeather}
\alias{ShinyWeather}
\title{A Child-Friendly Weather App}
\section{Shiny App}{

The \emph{ShinyWeather} app can be run using the runShinyWeather() function.

This App is meant to be used by parents together with children. It helps explain the weather to children in a simple way.
}

\section{The display contains}{

\itemize{
\item a \emph{map of the world centered on the Netherlands}
\item a \emph{calendar field}
\item \emph{values for temperature, rain, snow and wind}
\item \emph{progress bars showing temperature, rain, snow and wind}
\item a \emph{"Show Weather" button}
\item a \emph{"Show Activity" button}
\item an \emph{image placeholder for activities}
\item an \emph{image placeholder for clothes}
}
}

\section{Usage}{

First choose a location by clicking on the map and a date by selecting one in the calendar field.
It is possible to check the weather forecast for up to seven days including the current day.

By clicking "Show Weather" a field containing weather variables will appear.
The weather variables displayed are: \emph{temperature},  \emph{rain},\emph{snow} and \emph{wind speed}.

By clicking "Show Activities" a field containing images of activities will be displayed.
A text description will be shown under each image.
Browse through the activities using the "Back" and "Forward" buttons.
The activities displayed are based on the weather variables. They show what kind of activities can be done based on the weather conditions of the chosen day.
}

