% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshinyweather.R
\name{runshinyweather}
\alias{runshinyweather}
\title{runshinyweather}
\usage{
runshinyweather()
}
\description{
This function runs a shiny app that can be used to see the weather
forecast of a desired city.
}
