% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyweather.R
\docType{package}
\name{shinyweather}
\alias{shinyweather}
\alias{shinyweather-package}
\title{shinyweather}
\description{
This package contains a shiny app that can be used to see the weather
forecast of a desired city. To get the data, this package use the weatherforecat
package that is a R client for openweathermap API. You can see the details about
the weatherforecat package and about the openweathermap API here:
}
\references{
\url{https://github.com/nahfa911/weatherforecas}

\url{https://openweathermap.org/}
}
\author{
Nahid Farazmand
}
